+++ R/ffs_diff.R
@@ -1,10 +1,10 @@
 #' Estimation of bilateral migrant flows from bilateral migrant stocks using stock differencing approaches
 #' 
 #' Estimates migrant transitions flows between two sequential migrant stock tables using differencing approaches commonly used by economists.
-#' @param m1 Matrix of migrant stock totals at time \emph{t}. Rows in the matrix correspond to place of birth and columns to place of residence at time \emph{t} 
-#' @param m2 Matrix of migrant stock totals at time \emph{t}+1. Rows in the matrix correspond to place of birth and columns to place of residence at time \emph{t}+1.
+#' @param stock_start Matrix of migrant stock totals at time \emph{t}. Rows in the matrix correspond to place of birth and columns to place of residence at time \emph{t} 
+#' @param stock_end Matrix of migrant stock totals at time \emph{t}+1. Rows in the matrix correspond to place of birth and columns to place of residence at time \emph{t}+1.
 #' @param decrease How to treat decreases in bilateral stocks over the \emph{t} to \emph{t}+1 period (so as to avoid a negative bilateral flow estimates). See details for possible options. Default is \code{return}
-#' @param include_native_born Logical value to indicate whether to include diagonal elements of \code{m1} and \code{m2}. Default of \code{FALSE} - not include.
+#' @param include_native_born Logical value to indicate whether to include diagonal elements of \code{stock_start} and \code{stock_end}. Default of \code{FALSE} - not include.
 #'
 #' @return Estimates migrant transitions flows between two sequential migrant stock tables. 
 #' 
@@ -28,14 +28,14 @@
 #' dimnames(s1) <- dimnames(s2) <- list(pob = r, por = r)
 #' s1; s2
 #' 
-#' ffs_diff(m1 = s1, m2 = s2, decrease = "zero")
-#' ffs_diff(m1 = s1, m2 = s2, decrease = "return")
-ffs_diff <- function(m1, m2, decrease = "return", include_native_born = FALSE){
+#' ffs_diff(stock_start = s1, stock_end = s2, decrease = "zero")
+#' ffs_diff(stock_start = s1, stock_end = s2, decrease = "return")
+ffs_diff <- function(stock_start, stock_end, decrease = "return", include_native_born = FALSE){
   if(include_native_born == FALSE){
-    diag(m1) <- 0
-    diag(m2) <- 0
+    diag(stock_start) <- 0
+    diag(stock_end) <- 0
   }
-  y0 <- m2 - m1
+  y0 <- stock_end - stock_start
   if(decrease == "zero"){
     y0[y0<0] <- 0
     y <- y0
@@ -46,8 +46,8 @@ ffs_diff <- function(m1, m2, decrease = "return", include_native_born = FALSE){
     y0[y0<0] <- 0
     y <- y0 + y1
   }
-  if(!is.null(dimnames(m1))){
-    dimnames(y) <- list(orig = dimnames(m1)[[1]], dest = dimnames(m1)[[2]])
+  if(!is.null(dimnames(stock_start))){
+    dimnames(y) <- list(orig = dimnames(stock_start)[[1]], dest = dimnames(stock_start)[[2]])
   }
   return(y)
 }