+++ R/korea_dist.R
@@ -1,6 +1,0 @@
-#' Weighted distances in kilometers between 2020 population centroids of Korean administrative regions
-#'
-#' Distance matrix of kilometers between population weighted centroids in 2020 of first level administrative regions of South Korea.
-#' 
-#' @source Robin Edwards, Maksym Bondarenko, Andrew J. Tatem and Alessandro Sorichetta. Unconstrained subnational Population Weighted Density in 2000, 2005, 2010, 2015 and 2020 ( 100m resolution ). WorldPop, University of Southampton, UK. 
-"korea_dist"